🚚 refactor judge_core structure
