+++ lib/judge_core/src/lib.rs
@@ -1,10 +1,2 @@
-pub mod has_cycle;
-pub mod identifiers;
-pub mod job;
-pub mod judge_output;
-pub mod problem_registry;
-pub mod procedure;
-pub mod procedure_builder;
-pub mod registered_procedure_converter;
-pub mod runner;
-pub mod writer_schema_transpiler;
+pub mod logic;
+pub mod model;
