+++ lib/judge_core/src/logic/has_cycle.rs
