+++ lib/judge_core/src/model/procedure/writer_schema.rs
