+++ lib/problem_registry/src/client.rs
@@ -1,5 +1,5 @@
 #![allow(unused)]
-use judge_core::problem_registry;
+use judge_core::model::problem_registry;
 
 #[derive(Clone)]
 pub struct ProblemRegistryClient {}
@@ -8,7 +8,7 @@ pub struct ProblemRegistryClient {}
 impl problem_registry::ProblemRegistryClient for ProblemRegistryClient {
     async fn fetch(
         &self,
-        resource_id: judge_core::identifiers::ResourceId,
+        resource_id: judge_core::model::identifiers::ResourceId,
     ) -> Result<String, problem_registry::ResourceFetchError> {
         todo!()
     }
