+++ pylib/traopy/src/models/empty_directory.rs
@@ -1,4 +1,4 @@
-use judge_core::procedure::writer_schema::EmptyDirectory;
+use judge_core::model::procedure::writer_schema::EmptyDirectory;
 use pyo3::prelude::*;
 use pyo3_stub_gen::derive::*;
