+++ pylib/traopy/src/models/execution.rs
@@ -1,5 +1,5 @@
 use super::{dependency::*, output::PyScriptOutput};
-use judge_core::procedure::writer_schema::{Dependency, Execution};
+use judge_core::model::procedure::writer_schema::{Dependency, Execution};
 use pyo3::prelude::*;
 use pyo3_stub_gen::derive::*;
